+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## (IN PROGRESS)
 
+## [3.3.0](https://github.com/folio-org/ui-invoice/tree/v3.3.0) (2022-10-25)
+[Full Changelog](https://github.com/folio-org/ui-invoice/compare/v3.2.3...v3.3.0)
+
 * Add icons for invoice approval and pay actions. Refs UINV-426.
 * Display Vendor code rather than Vendor name. Refs UINV-427.
 * Upgrade `users` interface to `16.0`. Refs UINV-431.